Restructured personal details task tests
