--- application/tests/test_personal_details/test_views/test_manual_address.py
@@ -1,124 +1,0 @@
-from ..base_tests import PersonalDetailsTests, authenticate
-from django.urls import resolve, reverse
-from unittest import mock
-from application.presentation.personal_details.views import *
-import uuid
-from django.template.response import TemplateResponse
-
-from application.tests.test_utils import side_effect
-from application.services.db_gateways import NannyGatewayActions, IdentityGatewayActions
-
-
-@mock.patch.object(IdentityGatewayActions, "read", authenticate)
-class ManualEntryTests(PersonalDetailsTests):
-
-    def test_manual_entry_url_resolves_to_page(self):
-        found = resolve(reverse('personal-details:Personal-Details-Manual-Address'))
-        self.assertEqual(found.func.__name__, PersonalDetailManualAddressView.__name__)
-
-    def test_can_render_manual_entry_page(self):
-        """
-        Test to assert that the 'manual entry' page can be rendered.
-        """
-        with mock.patch.object(NannyGatewayActions, 'read') as nanny_api_read, \
-            mock.patch.object(NannyGatewayActions, 'list') as nanny_api_list,\
-            mock.patch.object(NannyGatewayActions, 'put') as nanny_api_put, \
-            mock.patch.object(NannyGatewayActions, 'delete') as nanny_api_delete, \
-            mock.patch.object(NannyGatewayActions, 'create') as nanny_api_create, \
-            mock.patch.object(NannyGatewayActions, 'patch') as nanny_api_patch, \
-            mock.patch.object(IdentityGatewayActions, 'read') as identity_api_read:
-
-            nanny_api_read.side_effect = side_effect
-            nanny_api_put.side_effect = side_effect
-
-            response = self.client.get(build_url('personal-details:Personal-Details-Manual-Address', get={
-                'id': uuid.UUID
-            }))
-
-            self.assertEqual(response.status_code, 200)
-
-    def test_can_update_address_valid_manual_entry_page(self):
-        """
-        Test to assert that the 'manual entry' page can be rendered.
-        """
-        with mock.patch.object(NannyGatewayActions, 'read') as nanny_api_read, \
-            mock.patch.object(NannyGatewayActions, 'list') as nanny_api_list,\
-            mock.patch.object(NannyGatewayActions, 'put') as nanny_api_put, \
-            mock.patch.object(NannyGatewayActions, 'delete') as nanny_api_delete, \
-            mock.patch.object(NannyGatewayActions, 'create') as nanny_api_create, \
-            mock.patch.object(NannyGatewayActions, 'patch') as nanny_api_patch, \
-            mock.patch.object(IdentityGatewayActions, 'read') as identity_api_read:
-
-            nanny_api_read.side_effect = side_effect
-            nanny_api_put.side_effect = side_effect
-
-            response = self.client.post(build_url('personal-details:Personal-Details-Manual-Address', get={
-                'id': uuid.UUID
-            }), {
-                'street_line1': 'Test',
-                'street_line2': '',
-                'town': 'test',
-                'county': '',
-                'postcode': 'WA14 4PA'
-            })
-
-            self.assertEqual(response.status_code, 302)
-            self.assertTrue('/home-address-details/' in response.url)
-
-    def test_can_create_address_valid_manual_entry_page(self):
-        """
-        Test to assert that the 'manual entry' page can be rendered.
-        """
-        with mock.patch.object(NannyGatewayActions, 'read') as nanny_api_read, \
-            mock.patch.object(NannyGatewayActions, 'list') as nanny_api_list,\
-            mock.patch.object(NannyGatewayActions, 'put') as nanny_api_put, \
-            mock.patch.object(NannyGatewayActions, 'delete') as nanny_api_delete, \
-            mock.patch.object(NannyGatewayActions, 'create') as nanny_api_create, \
-            mock.patch.object(NannyGatewayActions, 'patch') as nanny_api_patch, \
-            mock.patch.object(IdentityGatewayActions, 'read') as identity_api_read:
-
-            nanny_api_read.side_effect = side_effect
-            nanny_api_put.side_effect = side_effect
-
-            response = self.client.post(build_url('personal-details:Personal-Details-Manual-Address', get={
-                'id': uuid.UUID
-            }), {
-                'street_line1': 'Test',
-                'street_line2': '',
-                'town': 'test',
-                'county': '',
-                'postcode': 'WA14 4PA'
-            })
-
-            self.assertEqual(response.status_code, 302)
-            self.assertTrue('/home-address-details/' in response.url)
-
-    def test_can_submit_invalid_manual_entry_page(self):
-        """
-        Test to assert that the 'manual entry' page can be rendered.
-        """
-        with mock.patch.object(NannyGatewayActions, 'read') as nanny_api_read, \
-            mock.patch.object(NannyGatewayActions, 'list') as nanny_api_list,\
-            mock.patch.object(NannyGatewayActions, 'put') as nanny_api_put, \
-            mock.patch.object(NannyGatewayActions, 'delete') as nanny_api_delete, \
-            mock.patch.object(NannyGatewayActions, 'create') as nanny_api_create, \
-            mock.patch.object(NannyGatewayActions, 'patch') as nanny_api_patch, \
-            mock.patch.object(IdentityGatewayActions, 'read') as identity_api_read:
-
-            nanny_api_read.side_effect = side_effect
-            nanny_api_put.side_effect = side_effect
-
-            response = self.client.post(build_url('personal-details:Personal-Details-Manual-Address', get={
-                'id': uuid.UUID
-            }), {
-                'street_line1': '',
-                'street_line2': '',
-                'town': '',
-                'county': '',
-                'postcode': ''
-            })
-
-            self.assertEqual(response.status_code, 200)
-            self.assertTrue(type(response) == TemplateResponse)
-
-
